backend/requirements.txt
praw
langchain
python-dotenv
pydantic
langchain-community
langchain-google-genai
fastapi
uvicorn
